@@ -12,7 +12,7 @@ class AboutUsView extends StatelessWidget {
           Container(
             width: 300,
             margin: EdgeInsets.all(5),
-            child: RaisedButton(
+            child: ElevatedButton(
               child: Text(
                 'About The Application',
                 style: TextStyle(fontWeight: FontWeight.bold),
@@ -21,16 +21,18 @@ class AboutUsView extends StatelessWidget {
                 Navigator.push(context,
                     MaterialPageRoute(builder: (context) => AboutAppView()));
               },
-              color: Color(0xff4d74c2),
-              shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(15.0),
-                  side: BorderSide(color: Color(0xff285bc1))),
+              style: ElevatedButton.styleFrom(
+                shape: new RoundedRectangleBorder(
+                    borderRadius: new BorderRadius.circular(15.0),
+                    side: BorderSide(color: Color(0xff285bc1))),
+                primary: Color(0xff4d74c2),
+              ),
             ),
           ),
           Container(
             width: 300,
             margin: EdgeInsets.all(5),
-            child: RaisedButton(
+            child: ElevatedButton(
               child: Text(
                 'Contact Us',
                 style: TextStyle(fontWeight: FontWeight.bold),
@@ -39,16 +41,18 @@ class AboutUsView extends StatelessWidget {
                 Navigator.push(context,
                     MaterialPageRoute(builder: (context) => ContactUsView()));
               },
-              color: Color(0xff4d74c2),
-              shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(15.0),
-                  side: BorderSide(color: Color(0xff285bc1))),
+              style: ElevatedButton.styleFrom(
+                shape: new RoundedRectangleBorder(
+                    borderRadius: new BorderRadius.circular(15.0),
+                    side: BorderSide(color: Color(0xff285bc1))),
+                primary: Color(0xff4d74c2),
+              ),
             ),
           ),
           Container(
             width: 300,
             margin: EdgeInsets.all(5),
-            child: RaisedButton(
+            child: ElevatedButton(
               child: Text(
                 'About Us',
                 style: TextStyle(fontWeight: FontWeight.bold),
@@ -57,16 +61,18 @@ class AboutUsView extends StatelessWidget {
                 Navigator.push(context,
                     MaterialPageRoute(builder: (context) => AboutUsInfoView()));
               },
-              color: Color(0xff4d74c2),
-              shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(15.0),
-                  side: BorderSide(color: Color(0xff285bc1))),
+              style: ElevatedButton.styleFrom(
+                shape: new RoundedRectangleBorder(
+                    borderRadius: new BorderRadius.circular(15.0),
+                    side: BorderSide(color: Color(0xff285bc1))),
+                primary: Color(0xff4d74c2),
+              ),
             ),
           ),
           Container(
             width: 300,
             margin: EdgeInsets.all(5),
-            child: RaisedButton(
+            child: ElevatedButton(
               child: Text(
                 'Version',
                 style: TextStyle(fontWeight: FontWeight.bold),
@@ -75,10 +81,12 @@ class AboutUsView extends StatelessWidget {
                 Navigator.push(context,
                     MaterialPageRoute(builder: (context) => VersionView()));
               },
-              color: Color(0xff4d74c2),
-              shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(15.0),
-                  side: BorderSide(color: Color(0xff285bc1))),
+              style: ElevatedButton.styleFrom(
+                shape: new RoundedRectangleBorder(
+                    borderRadius: new BorderRadius.circular(15.0),
+                    side: BorderSide(color: Color(0xff285bc1))),
+                primary: Color(0xff4d74c2),
+              ),
             ),
           ),
         ],
