@@ -37,7 +37,7 @@ function UnsubscribeSuccessComponent(){
                                     </Button>
                                 </Link>
                                 
-                                <Button className="ml-4" style={{ textDecoration: 'none', color: '#FFF' }} href="https://github.com/clarkeustaquio/HackFest-2020-KMLNGMLKS-Corp" variant="contained" color="primary">
+                                <Button className="ml-4" style={{ textDecoration: 'none', color: '#FFF' }} href="https://github.com/jeikatsu/Bayanihan-News" variant="contained" color="primary">
                                     Download
                                 </Button>
                             </div>   
